+++ tests/ut/util-BrowserFileUtil.webtest.ts
@@ -1,46 +1,40 @@
-import {describe, expect, it} from 'vitest'
+import {beforeAll, afterAll, describe, expect, it} from 'vitest'
 import {
-  slice_file,
+  calc_crc64,
+  calc_file_crc64,
   calc_sha1,
   calc_sha256,
-  calc_crc64,
   calc_file_sha1,
   calc_file_sha256,
-  calc_file_crc64,
   calc_file_parts_sha1,
   calc_file_parts_sha256,
-  does_file_exist,
 } from '../../lib/context/BrowserFileUtil'
 import {init_chunks_sha} from '../../lib/utils/ChunkUtil'
 
-describe('src/context/BrowserFileUtil', () => {
-  describe('slice_file', () => {
-    it('slice_file', async () => {
-      let file = new File(['abc123456'], 'hello.png', {type: 'plain/text'})
-      let result = await slice_file(file, 0, 3)
-      expect(file.size).toBe(9)
-      expect(result.size).toBe(3)
+let ind = 0
+const type_arr = ['worker', 'wasm']
+
+for (let TYPE of type_arr) {
+  describe(`src/context/BrowserFileUtil [${TYPE}]`, () => {
+    beforeAll(() => {
+      ;(window as any).PDS_CALC_HASH_TYPE = type_arr[ind]
     })
-  })
-  describe('calc_crc64', () => {
-    it('calc_crc64', () => {
-      expect(calc_crc64('abc', '0')).toBe('3231342946509354535')
-      expect(calc_crc64('中文')).toBe('8230427039312370437')
+    afterAll(() => {
+      ind++
     })
-  })
-  describe('calc_sha1', () => {
-    it('calc_sha1', () => {
-      expect(calc_sha1('abc')).toBe('A9993E364706816ABA3E25717850C26C9CD0D89D')
-      expect(calc_sha1('中文')).toBe('7BE2D2D20C106EEE0836C9BC2B939890A78E8FB3')
+
+    it(`calc_crc64`, async () => {
+      expect(await calc_crc64('abc', '0')).toBe('3231342946509354535')
+      expect(await calc_crc64('中文', undefined)).toBe('8230427039312370437')
     })
-  })
-  describe('calc_sha256', () => {
-    it('calc_sha256', () => {
-      expect(calc_sha256('abc')).toBe('BA7816BF8F01CFEA414140DE5DAE2223B00361A396177A9CB410FF61F20015AD')
-      expect(calc_sha256('中文')).toBe('72726D8818F693066CEB69AFA364218B692E62EA92B385782363780F47529C21')
+    it('calc_sha1', async () => {
+      expect(await calc_sha1('abc')).toBe('A9993E364706816ABA3E25717850C26C9CD0D89D')
+      expect(await calc_sha1('中文')).toBe('7BE2D2D20C106EEE0836C9BC2B939890A78E8FB3')
     })
-  })
-  describe('calc_file_sha1', () => {
+    it('calc_sha256', async () => {
+      expect(await calc_sha256('abc')).toBe('BA7816BF8F01CFEA414140DE5DAE2223B00361A396177A9CB410FF61F20015AD')
+      expect(await calc_sha256('中文')).toBe('72726D8818F693066CEB69AFA364218B692E62EA92B385782363780F47529C21')
+    })
     it('calc_file_sha1', async () => {
       let file = new File(['abc'], 'hello.png', {type: 'plain/text'})
       let prog = 0
@@ -57,8 +51,6 @@ describe('src/context/BrowserFileUtil', () => {
       expect(result).toBe('A9993E364706816ABA3E25717850C26C9CD0D89D')
       expect(prog).toBe(100)
     })
-  })
-  describe('calc_file_sha256', () => {
     it('calc_file_sha256', async () => {
       let file = new File(['abc'], 'hello.png', {type: 'plain/text'})
       let prog = 0
@@ -75,8 +67,6 @@ describe('src/context/BrowserFileUtil', () => {
       expect(result).toBe('BA7816BF8F01CFEA414140DE5DAE2223B00361A396177A9CB410FF61F20015AD')
       expect(prog).toBe(100)
     })
-  })
-  describe('calc_file_parts_sha1', () => {
     it('calc_file_parts_sha1', async () => {
       let t: string[] = []
       for (let i = 0; i < 100; i++) t.push(`test-${i}`)
@@ -114,8 +104,6 @@ describe('src/context/BrowserFileUtil', () => {
 
       expect(prog).toBe(100)
     })
-  })
-  describe('calc_file_parts_sha256', () => {
     it('calc_file_parts_sha256', async () => {
       let t: string[] = []
       for (let i = 0; i < 100; i++) t.push(`test-${i}`)
@@ -154,9 +142,8 @@ describe('src/context/BrowserFileUtil', () => {
 
       expect(prog).toBe(100)
     })
-  })
-  describe('calc_file_crc64', () => {
-    it('file', async () => {
+
+    it('calc_file_crc64 file', async () => {
       let t: string[] = []
       for (let i = 0; i < 100; i++) t.push(`test-${i}`)
 
@@ -173,7 +160,7 @@ describe('src/context/BrowserFileUtil', () => {
       expect(prog).toBe(100)
       expect(last).toBe('4762498020522648846')
     })
-    it('empty file', async () => {
+    it('calc_file_crc64 empty file', async () => {
       const file = new File([], 'hello.txt', {type: 'plain/text'})
       let prog = 0
       let last = await calc_file_crc64(
@@ -188,25 +175,4 @@ describe('src/context/BrowserFileUtil', () => {
       expect(prog).toBe(100)
     })
   })
-
-  describe('does_file_exist', () => {
-    it('file', async () => {
-      let t: string[] = []
-      for (let i = 0; i < 100; i++) t.push(`test-${i}`)
-
-      const file = new File([t.join('\n')], 'hello.txt', {type: 'plain/text'})
-
-      let err = await does_file_exist(file)
-      expect(err).toBe(undefined)
-    })
-    it('empty file', async () => {
-      let t: string[] = []
-      for (let i = 0; i < 100; i++) t.push(`test-${i}`)
-
-      const file = new File([], 'hello.txt', {type: 'plain/text'})
-
-      let err = await does_file_exist(file)
-      expect(err).toBe(undefined)
-    })
-  })
-})
+}