+++ vite.config.browser.js
@@ -32,7 +32,8 @@ export default defineConfig({
       reportsDirectory: join(__dirname, './coverage/browser'),
       include: ['lib/**/*.ts', 'lib/**/*.js'],
       exclude: [
-        '**/*-wasm.js',
+        'lib/utils/crc64/wasm/crc64-wasm.js',
+        'lib/utils/sha-hash/wasm/wasm.js',
         'tests',
         'lib/utils/crc64/wasm/index.js',
         'lib/utils/axios-node-adapter/*',